NARE is now visualized as an n-tree, respecting the specification pattern in the DSL
At its core, the results are still calculated as a binary tree.
However, the intermediate nodes created by the DSL is now discarded, leaving only those nodes with a identity in the resulting specification and evalation tree.
This change simplifies the Influx/Prometheus querying, as well as it makes the visualization of the evaluation much closer to the specification DSL